refactor(category): Category 생성 로직을 CommandService로 이동
- 기존 CategoryQueryService에 있던 create 메서드를 CategoryCommandService로 분리
